chore: moved benchmarks to a subpackage so they can be run in isolation
